save edited method

+++ app/src/main/java/com/hgshkt/todolist/ItemAdapter.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
 import android.widget.EditText
 import android.widget.ImageView
 import android.widget.TextView
@@ -19,8 +18,31 @@ class ItemAdapter(private val context: Context, private val items: List<Item>,
     private val dao: ItemDao)
     : RecyclerView.Adapter<ItemAdapter.ImageViewHolder>() {
 
-    // current item which is being edited
-    var editPosition: Int? = null
+
+    companion object {
+        // current item which is being edited
+        var editPosition: Int? = null
+
+        fun saveEdited() {
+            val lastEditedItem = MainActivity.recyclerView.layoutManager!!.findViewByPosition(editPosition!!)
+            val editTitle = lastEditedItem!!.findViewById<EditText>(R.id.editTitle)
+            val itemTitle = lastEditedItem.findViewById<TextView>(R.id.itemTitle)
+            val saveButton = lastEditedItem.findViewById<ImageView>(R.id.saveButton)
+            val editButton = lastEditedItem.findViewById<ImageView>(R.id.editButton)
+
+            itemTitle.text = editTitle.text
+            editTitle.visibility = View.INVISIBLE
+            itemTitle.visibility = View.VISIBLE
+
+            saveButton.visibility = View.INVISIBLE
+            editButton.visibility = View.VISIBLE
+
+            db.getItemDao().updateItem(Item(
+                editTitle.text.toString()
+            ))
+            editPosition = null
+        }
+    }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ImageViewHolder {
         val view = LayoutInflater.from(context).inflate(R.layout.item, parent, false)
@@ -49,75 +71,54 @@ class ItemAdapter(private val context: Context, private val items: List<Item>,
 
         holder.editButton.setOnClickListener {
             if (editPosition != null) {
-                val lastEditedItem = MainActivity.recyclerView.layoutManager!!.findViewByPosition(editPosition!!)
-                val editTitle = lastEditedItem!!.findViewById<EditText>(R.id.editTitle)
-                val itemTitle = lastEditedItem.findViewById<EditText>(R.id.itemTitle)
-                val saveButton = lastEditedItem.findViewById<EditText>(R.id.saveButton)
-                val editButton = lastEditedItem.findViewById<EditText>(R.id.editButton)
+                saveEdited()
+            } else {
+                holder.edit.setText(holder.title.text)
+                holder.title.visibility = View.INVISIBLE
+                holder.edit.visibility = View.VISIBLE
 
-                itemTitle.text = editTitle.text
-                editTitle.visibility = View.INVISIBLE
-                itemTitle.visibility = View.VISIBLE
+                holder.editButton.visibility = View.INVISIBLE
+                holder.saveButton.visibility = View.VISIBLE
 
-                saveButton.visibility = View.INVISIBLE
-                editButton.visibility = View.VISIBLE
-
-                db.getItemDao().updateItem(Item(
-                    editTitle.text.toString()
-                ))
+                editPosition = position
             }
-
-            holder.edit.setText(holder.title.text)
-            holder.title.visibility = View.INVISIBLE
-            holder.edit.visibility = View.VISIBLE
-
-            holder.editButton.visibility = View.INVISIBLE
-            holder.saveButton.visibility = View.VISIBLE
-
-            editPosition = position
         }
 
         holder.deleteButton.setOnClickListener {
-            dao.delete(currentItem)
-            (context as MainActivity).update()
+            if (editPosition != null) {
+                saveEdited()
+            } else {
+                dao.delete(currentItem)
+                (context as MainActivity).update()
+            }
         }
 
         holder.tick.setOnClickListener {
-            currentItem.complete = !currentItem.complete
-            loadTick(currentItem.complete, holder.tick)
-            dao.updateItem(currentItem)
-            (context as MainActivity).update()
+            if (editPosition != null) {
+                saveEdited()
+            } else {
+                currentItem.complete = !currentItem.complete
+                loadTick(currentItem.complete, holder.tick)
+                dao.updateItem(currentItem)
+                (context as MainActivity).update()
+            }
         }
 
         holder.title.setOnClickListener {
             //  if user has not saved the changes in previous item, save will happen automatically
             if (editPosition != null) {
-                val lastEditedItem = MainActivity.recyclerView.layoutManager!!.findViewByPosition(editPosition!!)
-                val editTitle = lastEditedItem!!.findViewById<EditText>(R.id.editTitle)
-                val itemTitle = lastEditedItem.findViewById<TextView>(R.id.itemTitle)
-                val saveButton = lastEditedItem.findViewById<ImageView>(R.id.saveButton)
-                val editButton = lastEditedItem.findViewById<ImageView>(R.id.editButton)
+                saveEdited()
+            } else {
 
-                itemTitle.text = editTitle.text
-                editTitle.visibility = View.INVISIBLE
-                itemTitle.visibility = View.VISIBLE
+                holder.edit.setText(holder.title.text)
+                holder.title.visibility = View.INVISIBLE
+                holder.edit.visibility = View.VISIBLE
 
-                saveButton.visibility = View.INVISIBLE
-                editButton.visibility = View.VISIBLE
+                holder.editButton.visibility = View.INVISIBLE
+                holder.saveButton.visibility = View.VISIBLE
 
-                db.getItemDao().updateItem(Item(
-                    editTitle.text.toString()
-                ))
+                editPosition = position
             }
-
-            holder.edit.setText(holder.title.text)
-            holder.title.visibility = View.INVISIBLE
-            holder.edit.visibility = View.VISIBLE
-
-            holder.editButton.visibility = View.INVISIBLE
-            holder.saveButton.visibility = View.VISIBLE
-
-            editPosition = position
         }
     }
